test: add L3 networks nodeport ETP: Cluster test for GDC-SO
Bug: b/390022320
Carry-Patch-Exempt: Test Infra Changes
Backport-Approval-Exempt: Test Infra Changes
Reviewed-on: https://gke-internal-review.googlesource.com/c/third_party/cilium/+/1300286
Prow-Auto-Submit: Shouri Piratla <shouri@google.com>
Unit-Verified: Prow_Bot_V2 <425329972751-compute@developer.gserviceaccount.com>
Tested-by: Prow_Bot_V2 <425329972751-compute@developer.gserviceaccount.com>
Louhi-Verified: Louhi_Bot <louhi-gerrit-proxy@google.com>
Lint: Lint 🤖 <android-build-ayeaye@system.gserviceaccount.com>
Reviewed-on: https://gke-internal-review.googlesource.com/c/third_party/cilium/+/1392323

+++ google_test/wora/pkg/test/utils/utils.go
@@ -361,6 +361,9 @@ func cleanupResources(ctx context.Context, cl k8sclient.Client, podName, namespa
 
 	// Delete associated network interfaces
 	for _, info := range networkInfos {
+		if info.IPAMMode == networkv1.InternalMode {
+			continue
+		}
 		niName := fmt.Sprintf("%s-%s", podName, info.InterfaceName)
 		err := cl.Delete(ctx, &networkv1.NetworkInterface{
 			ObjectMeta: metav1.ObjectMeta{
